+++ index.js
@@ -37,8 +37,9 @@ module.exports = function (options) {
         throw new PluginError(PLUGIN_NAME, 'For packages <> "$TMP" a transport number is necessary.');
     }
 
-    if (options.ui5.bspcontainer.length > 15) {
-        throw new PluginError(PLUGIN_NAME, '"ui5.bspcontainer" option must not be longer than 15 characters.');
+    var bspcontainerExclNamespace = options.ui5.bspcontainer.substring(options.ui5.bspcontainer.lastIndexOf('/') + 1);
+    if (bspcontainerExclNamespace > 15) {
+        throw new PluginError(PLUGIN_NAME, '"ui5.bspcontainer" option must not be longer than 15 characters (exclusive customer specific namespace e.g. /YYY/..');
     }
 
     if (!options.ui5.language) {