+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixes
 - \#2 Provide Language 
+- Restrict BPS-Container length to 15 characters without namespace
 
 ## 0.1.1 (2016-06-12)